+++ karma.config.js
@@ -1,42 +1,21 @@
-// Karma configuration for LaxarJS core
+/**
+ * Copyright 2017 aixigo AG
+ * Released under the MIT license.
+ * http://laxarjs.org/license
+ */
 /* eslint-env node */
-
-const webpackConfig = Object.assign( {}, require('./webpack.config' ) );
-delete webpackConfig.entry;
-delete webpackConfig.plugins;
-webpackConfig.devtool = 'inline-source-map';
+const pkg = require( './package.json' );
+const laxarInfrastructure = require( 'laxar-infrastructure' );
 
 module.exports = function( config ) {
-   const browsers = [ 'PhantomJS', 'Firefox' ].concat( [
-      process.env.TRAVIS ? 'ChromeTravisCi' : 'Chrome'
-   ] );
-
-   config.set( {
-      frameworks: [ 'jasmine' ],
-      files: [
-         require.resolve( 'laxar/dist/polyfills' ),
-         'spec/*.spec.js'
-      ],
-      preprocessors: {
-         'spec/*.spec.js': [ 'webpack', 'sourcemap' ]
-      },
-      webpack: webpackConfig,
-
-      reporters: [ 'progress', 'junit' ],
-      junitReporter: {
-         outputDir: 'karma-output/'
-      },
-      port: 9876,
-      browsers,
-      customLaunchers: {
-         ChromeTravisCi: {
-            base: 'Chrome',
-            flags: [ '--no-sandbox' ]
-         }
-      },
-      browserNoActivityTimeout: 100000,
-      singleRun: true,
-      autoWatch: false,
-      concurrency: Infinity
-   } );
+   config.set( karmaConfig() );
 };
+
+function karmaConfig() {
+   const webpackBaseConfig = require( './webpack.config' )[ 0 ];
+   return laxarInfrastructure.karma( [ `spec/${pkg.name}.spec.js` ], {
+      context: __dirname,
+      resolve: webpackBaseConfig.resolve,
+      module: webpackBaseConfig.module
+   } );
+}