+++ webpack.config.js
@@ -6,22 +6,10 @@
 /* eslint-env node */
 
 const path = require( 'path' );
-const webpack = require( 'webpack' );
+const pkg = require( './package.json' );
 
-const nodeEnv = process.env.NODE_ENV;
-const isProduction = nodeEnv === 'production';
-const isBrowserSpec = nodeEnv === 'browser-spec';
-
-const name = require( './package.json' ).name;
-const externals = {
-   'laxar': 'laxar',
-   'laxar-tooling': 'laxar-tooling',
-   'laxar/laxar-widget-service-mocks': 'laxar/dist/laxar-widget-service-mocks'
-};
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-const baseConfig = {
+const webpack = require( 'laxar-infrastructure' ).webpack( {
+   context: __dirname,
    resolve: {
       // 'webpack' is included to get rid of ajv warnings
       mainFields: [ 'webpack', 'browser', 'main', 'index' ]
@@ -30,73 +18,19 @@ const baseConfig = {
       rules: [
          {
             test: /\.js$/,
-            exclude: /(node_modules)\/(?!laxar.*)/,
+            include: [
+               path.resolve( __dirname, pkg.main ),
+               path.resolve( __dirname, 'lib/' ),
+               path.resolve( __dirname, 'spec/' ),
+               path.resolve( __dirname, 'node_modules/laxar/' )
+            ],
             loader: 'babel-loader'
          }
       ]
    }
-};
+} );
 
-const config = isProduction ? distConfig() : baseConfig;
-
-if( isBrowserSpec ) {
-   const WebpackJasmineHtmlRunnerPlugin = require( 'webpack-jasmine-html-runner-plugin' );
-   config.entry = WebpackJasmineHtmlRunnerPlugin.entry( './spec/*.spec.js' );
-   config.plugins = [ new WebpackJasmineHtmlRunnerPlugin() ];
-   config.output = {
-      path: path.resolve( path.join( process.cwd(), 'spec-output' ) ),
-      publicPath: '/spec-output/',
-      filename: '[name].bundle.js'
-   };
-}
-
-module.exports = config;
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-function distConfig() {
-
-   return [
-      distConfigItem( `./${name}.js`, `./${name}.js` ),
-      distConfigItem( `./${name}.js`, `./${name}.min.js`, { minify: true } )
-   ];
-
-   function distConfigItem( entry, output, optionalOptions ) {
-      const options = Object.assign( {
-         minify: false,
-         externals
-      }, optionalOptions || {} );
-
-      const config = Object.assign( {}, baseConfig );
-
-      config.entry = entry;
-
-      config.output = {
-         path: path.resolve( __dirname ),
-         filename: `dist/${output}`,
-         library: name,
-         libraryTarget: 'umd',
-         umdNamedDefine: true
-      };
-
-      config.externals = options.externals;
-
-      config.plugins = [
-         new webpack.SourceMapDevToolPlugin( {
-            filename: `dist/${output}.map`
-         } )
-      ];
-
-      if( options.minify ) {
-         config.plugins.push(
-            new webpack.optimize.UglifyJsPlugin( {
-               compress: { warnings: false },
-               sourceMap: true
-            } )
-         );
-      }
-
-      return config;
-   }
-
-}
+module.exports = [
+   webpack.library(),
+   webpack.browserSpec( [ `./spec/${pkg.name}.spec.js` ] )
+];